Revert "1564"
This reverts commit 9afe6f9c064a02da6b18227157a30c468de43c55.

+++ src/playground_base/playground_base/base/topo/param/FBLinearParam.hpp
@@ -13,8 +13,8 @@ struct FBLinearParamRealTime
 {
   float min = 0.0f;
   float max = 1.0f;
-  double editSkewFactor = 1.0f;
-  double displayMultiplier = 1.0f;
+  float editSkewFactor = 1.0f;
+  float displayMultiplier = 1.0f;
 
   FB_NOCOPY_NOMOVE_DEFCTOR(FBLinearParamRealTime);
   float NormalizedToPlain(float normalized) const;
