+++ src/playground_base/playground_base/gui/controls/FBGUIParamSlider.hpp
@@ -24,5 +24,5 @@ public:
   juce::String getTextFromValue(double value) override;
   double getValueFromText(const juce::String& text) override;
 
-  void SetValueNormalizedFromPlug(double normalized) override;
+  void SetValueNormalizedFromPlug(float normalized) override;
 };
